+++ src/App.test.js
@@ -1,30 +1,32 @@
 import { render, screen } from "@testing-library/react";
-import { Timer } from "./App";
+import { Timer } from "./Timer";
 import { checkWord, Status } from "./checkWord";
 import timekeeper from "timekeeper";
 
-test("victory timer works at midnight", () => {
-  timekeeper.freeze(new Date("2021-01-16T00:00:00.000Z"));
-  render(<Timer targetDate="2021-01-16T00:00:00.000Z" />);
-  const linkElement = screen.getByText(/00:00:00/i);
-  expect(linkElement).toBeInTheDocument();
-  timekeeper.reset();
-});
+describe("Victory Timer", () => {
+  test("works at midnight", () => {
+    timekeeper.freeze(new Date("2021-01-16T00:00:00.000Z"));
+    render(<Timer targetDate="2021-01-16T00:00:00.000Z" />);
+    const linkElement = screen.getByText(/00:00:00/i);
+    expect(linkElement).toBeInTheDocument();
+    timekeeper.reset();
+  });
 
-test("victory timer works before midnight", () => {
-  timekeeper.freeze(new Date("2021-01-15T23:00:00.000Z"));
-  render(<Timer targetDate="2021-01-16T00:00:00.000Z" />);
-  const linkElement = screen.getByText(/01:00:00/i);
-  expect(linkElement).toBeInTheDocument();
-  timekeeper.reset();
-});
+  test("works before midnight", () => {
+    timekeeper.freeze(new Date("2021-01-15T23:00:00.000Z"));
+    render(<Timer targetDate="2021-01-16T00:00:00.000Z" />);
+    const linkElement = screen.getByText(/01:00:00/i);
+    expect(linkElement).toBeInTheDocument();
+    timekeeper.reset();
+  });
 
-test("victory timer works for complex time before midnight", () => {
-  timekeeper.freeze(new Date("2021-01-15T23:15:01.000Z"));
-  render(<Timer targetDate="2021-01-16T00:00:00.000Z" />);
-  const linkElement = screen.getByText(/00:44:59/i);
-  expect(linkElement).toBeInTheDocument();
-  timekeeper.reset();
+  test("works for complex time before midnight", () => {
+    timekeeper.freeze(new Date("2021-01-15T23:15:01.000Z"));
+    render(<Timer targetDate="2021-01-16T00:00:00.000Z" />);
+    const linkElement = screen.getByText(/00:44:59/i);
+    expect(linkElement).toBeInTheDocument();
+    timekeeper.reset();
+  });
 });
 
 describe("word checker", () => {
@@ -32,22 +34,22 @@ describe("word checker", () => {
     const result = checkWord("hello", "hello");
 
     expect(result).toEqual([
-      {letter: 'h', status: Status.CORRECT},
-      {letter: 'e', status: Status.CORRECT},
-      {letter: 'l', status: Status.CORRECT},
-      {letter: 'l', status: Status.CORRECT},
-      {letter: 'o', status: Status.CORRECT},
+      { letter: "h", status: Status.CORRECT },
+      { letter: "e", status: Status.CORRECT },
+      { letter: "l", status: Status.CORRECT },
+      { letter: "l", status: Status.CORRECT },
+      { letter: "o", status: Status.CORRECT },
     ]);
   });
   test("detects totally incorrect words", () => {
     const result = checkWord("hello", "sinks");
 
     expect(result).toEqual([
-      {letter: 'h', status: Status.INCORRECT},
-      {letter: 'e', status: Status.INCORRECT},
-      {letter: 'l', status: Status.INCORRECT},
-      {letter: 'l', status: Status.INCORRECT},
-      {letter: 'o', status: Status.INCORRECT},
+      { letter: "h", status: Status.INCORRECT },
+      { letter: "e", status: Status.INCORRECT },
+      { letter: "l", status: Status.INCORRECT },
+      { letter: "l", status: Status.INCORRECT },
+      { letter: "o", status: Status.INCORRECT },
     ]);
   });
 
@@ -55,11 +57,11 @@ describe("word checker", () => {
     const result = checkWord("hello", "ghost");
 
     expect(result).toEqual([
-      {letter: 'h', status: Status.PARTIAL},
-      {letter: 'e', status: Status.INCORRECT},
-      {letter: 'l', status: Status.INCORRECT},
-      {letter: 'l', status: Status.INCORRECT},
-      {letter: 'o', status: Status.PARTIAL},
+      { letter: "h", status: Status.PARTIAL },
+      { letter: "e", status: Status.INCORRECT },
+      { letter: "l", status: Status.INCORRECT },
+      { letter: "l", status: Status.INCORRECT },
+      { letter: "o", status: Status.PARTIAL },
     ]);
   });
 
@@ -67,11 +69,11 @@ describe("word checker", () => {
     const result = checkWord("share", "shire");
 
     expect(result).toEqual([
-      {letter: 's', status: Status.CORRECT},
-      {letter: 'h', status: Status.CORRECT},
-      {letter: 'a', status: Status.INCORRECT},
-      {letter: 'r', status: Status.CORRECT},
-      {letter: 'e', status: Status.CORRECT},
+      { letter: "s", status: Status.CORRECT },
+      { letter: "h", status: Status.CORRECT },
+      { letter: "a", status: Status.INCORRECT },
+      { letter: "r", status: Status.CORRECT },
+      { letter: "e", status: Status.CORRECT },
     ]);
   });
 
@@ -79,11 +81,11 @@ describe("word checker", () => {
     const result = checkWord("hello", "plane");
 
     expect(result).toEqual([
-      {letter: 'h', status: Status.INCORRECT},
-      {letter: 'e', status: Status.PARTIAL},
-      {letter: 'l', status: Status.PARTIAL},
-      {letter: 'l', status: Status.INCORRECT},
-      {letter: 'o', status: Status.INCORRECT},
+      { letter: "h", status: Status.INCORRECT },
+      { letter: "e", status: Status.PARTIAL },
+      { letter: "l", status: Status.PARTIAL },
+      { letter: "l", status: Status.INCORRECT },
+      { letter: "o", status: Status.INCORRECT },
     ]);
   });
 });
