+++ src/Timer.js
@@ -1,0 +1,39 @@
+import { useEffect, useState } from "react";
+
+
+export function Timer({ targetDate }) {
+  const calculateTimeLeft = () => {
+    let difference = +new Date(targetDate) - +new Date();
+    let timeLeft = {
+      hours: "00",
+      minutes: "00",
+      seconds: "00",
+    };
+
+    if (difference > 0) {
+      timeLeft = {
+        hours: Math.floor((difference / (1000 * 60 * 60)) % 24),
+        minutes: Math.floor((difference / 1000 / 60) % 60),
+        seconds: Math.floor((difference / 1000) % 60),
+      };
+    }
+
+    return timeLeft;
+  };
+
+  const [timeLeft, setTimeLeft] = useState(calculateTimeLeft());
+  useEffect(() => {
+    const timer = setTimeout(() => {
+      setTimeLeft(calculateTimeLeft());
+    }, 1000);
+    return () => clearTimeout(timer);
+  });
+
+  return (
+    <p>
+      Next puzzle in {String(timeLeft.hours).padStart(2, "0")}:
+      {String(timeLeft.minutes).padStart(2, "0")}:
+      {String(timeLeft.seconds).padStart(2, "0")}
+    </p>
+  );
+}
